Update Iteration3
RNN - LSTM forecasting is now working :
It is in a function called on start up and then saved so it is not called again while the application is open:
An improvement on this: instead of running every time the application is open, try every day: e.g. by if it isnt done that day, save to csv file. It is is already done, call csv file.